requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
ase=3.19.1=pypi_0
backcall=0.1.0=py38_0
blas=1.0=mkl
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py38_0
chardet=3.0.4=pypi_0
cloudpickle=1.3.0=py_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=pypi_0
cytoolz=0.10.1=py38h7b6447c_0
dask-core=2.14.0=py_0
dbus=1.13.12=h746ee38_0
decorator=4.4.2=py_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
glib=2.63.1=h5a9c865_0
googledrivedownloader=0.4=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.10.0=pypi_0
icu=58.2=h9c2bf20_1
idna=2.9=pypi_0
imageio=2.8.0=py_0
intel-openmp=2020.0=166
ipdb=0.13.2=pypi_0
ipython=7.13.0=py38h5ca1d4c_0
ipython_genutils=0.2.0=py38_0
isodate=0.6.0=pypi_0
jedi=0.16.0=py38_1
joblib=0.14.1=py_0
jpeg=9b=h024ee3a_2
kiwisolver=1.2.0=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
llvmlite=0.32.0=pypi_0
matplotlib=3.2.1=pypi_0
matplotlib-base=3.1.3=py38hef1b27d_0
mkl=2020.0=166
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.0.15=py38ha843d7b_0
mkl_random=1.1.0=py38h962f231_0
ncurses=6.2=he6710b0_0
networkx=2.4=py_0
ninja=1.9.0=py38hfd86e86_0
numba=0.49.0=pypi_0
numpy=1.18.1=py38h4f9e942_0
numpy-base=1.18.1=py38hde5b4d6_1
olefile=0.46=py_0
openssl=1.1.1f=h7b6447c_0
pandas=1.0.3=pypi_0
parso=0.6.2=py_0
pcre=8.43=he6710b0_0
pexpect=4.8.0=py38_0
pickleshare=0.7.5=py38_1000
pillow=7.0.0=py38hb39fc2d_0
pip=20.0.2=py38_1
plyfile=0.7.2=pypi_0
prompt-toolkit=3.0.4=py_0
prompt_toolkit=3.0.4=0
ptyprocess=0.6.0=py38_0
pygments=2.6.1=py_0
pyparsing=2.4.7=pypi_0
pyqt=5.9.2=py38h05f1152_4
python=3.8.2=hcf32534_0
python-dateutil=2.8.1=py_0
pytorch=1.4.0=py3.8_cuda10.1.243_cudnn7.6.3_0
pytz=2019.3=pypi_0
pywavelets=1.1.1=py38h7b6447c_0
qt=5.9.7=h5867ecd_1
rdflib=4.2.2=pypi_0
readline=8.0=h7b6447c_0
requests=2.23.0=pypi_0
scikit-image=0.16.2=py38h0573a6f_0
scikit-learn=0.22.1=py38hd81dba3_0
scipy=1.4.1=py38h0b6359f_0
setuptools=46.1.3=py38_0
sip=4.19.13=py38he6710b0_0
six=1.14.0=py38_0
sqlite=3.31.1=h7b6447c_0
tk=8.6.8=hbc83047_0
toolz=0.10.0=py_0
torch-cluster=1.5.4=pypi_0
torch-geometric=1.5.0=pypi_0
torch-scatter=2.0.4=pypi_0
torch-sparse=0.6.1=pypi_0
torch-spline-conv=1.2.0=pypi_0
torchvision=0.5.0=py38_cu101
tornado=6.0.4=py38h7b6447c_1
tqdm=4.44.1=py_0
traitlets=4.3.3=py38_0
urllib3=1.25.9=pypi_0
wcwidth=0.1.9=py_0
wheel=0.34.2=py38_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
